Update the check_reisstor_value() function, and add the decision to the main function.
check_resistor_value() function was with a simple list of lists, and right now it is
a lsit of dictionaries which males easier to search through the keys.
I solved the resistor_value_status problem, where it always send False status,
the problem was the next: I had an else into the for loop, so no matter the if found
the result, it continues to assign a False that was in the else statement.